Add OTel tracing bucket
Refactor client.NewBucket

Add client.NewInstrumentedBucket
