--- client/testconf/blank-gcs.conf.yml
@@ -1,1 +1,0 @@
-type: GCS
