--- client/testconf/fake-gcs.conf.yml
@@ -1,3 +1,0 @@
-type: FAKE-GCS
-config:
-  bucket: test-bucket
