+++ client/testconf/filesystem.conf.yml
@@ -1,0 +1,3 @@
+type: "FILESYSTEM"
+config:
+  directory: "./data"
